// src/interface-adapter/interfaces/reports/member/member-report.response.interface.ts
export interface IReportDataMemberResponse {
  kode_member: string;
  no_identitas: string;
  nama_lengkap: string;
  agama?: string;
  penghasilan?: string;
  pengeluaran?: string;
  jumlah_anak_pria?: number;
  jumlah_anak_wanita?: number;
  no_hp: string;
  telp_rumah?: string;
  telp_kantor?: string;
}